Elias TCError now takes Environment instead of Backtrace
